Pairs matrix

Compute the pairs matrix for a full ranking of m objects

Pmaker(m)

Arguments

m
the number of objects

Value

...

Details

This is the transpose of the pairs matrix presented in Marden (1995).

References

Marden, J. I. (1995). Analyzing and Modeling Rank Data, London: Chapman & Hall. p.42.

See also

Examples

data(city); city

#>       city suburb country
#> 1 2 3  210     22      10
#> 1 3 2   23      4       1
#> 2 1 3  111     45      14
#> 2 3 1    8      4       0
#> 3 1 2  204    299     125
#> 3 2 1   81    126     152

Pmaker(3)

#>     123 132 213 231 312 321
#> 1>2   1   1   0   1   0   0
#> 1>3   1   1   1   0   0   0
#> 2>3   1   0   1   0   1   0

Pmaker(3) %*% city

#>     city suburb country
#> 1>2  241     30      11
#> 1>3  344     71      25
#> 2>3  525    366     149

# 1 = city, 2 = suburb, 3 = country

# looking just among city folk, generate the pairs matrix
city[,"city",drop=FALSE] # the data

#>       city
#> 1 2 3  210
#> 1 3 2   23
#> 2 1 3  111
#> 2 3 1    8
#> 3 1 2  204
#> 3 2 1   81

m <- sum(city[,"city"])
k <- (Pmaker(3) %*% city)[,1]
Khat <- upper(k) + lower(m-k)
colnames(Khat) <- row.names(Khat) <- colnames(city)
Khat

#>         city suburb country
#> city       0    241     344
#> suburb   396      0     525
#> country  293    112       0

round(Khat / m, 2) # % times row is rated over column

#>         city suburb country
#> city    0.00   0.38    0.54
#> suburb  0.62   0.00    0.82
#> country 0.46   0.18    0.00

# worked out: city is voted over suburb in 123 , 132, and 231, equaling
210 + 23 + 8   # = Khat[1,2]

#> [1] 241

# whereas suburb is rated over city in 213, 312, 321, equaling
111 + 204 + 81 # = Khat[2,1]

#> [1] 396

# is there a condorcet choice?

p <- ncol(Khat)
Khat[which(diag(p) == 1)] <- NA
K2 <- t(apply(Khat, 1, function(v) v[!is.na(v)])) # remove diag elts
boole <- apply(K2/m, 1, function(x) all(x > .5))
if(any(boole)) names(boole)[which(boole)]

#> [1] "suburb"

# suburb is a condorcet choice