Means matrix (rank data)

Compute the means matrix for a full ranking of m objects

Smaker(m)

Arguments

m
the number of objects

Value

...

Details

This is the transpose of the means matrix presented in Marden (1995); it projects onto the means subspace of a collection of ranked data. See the examples for how to compute the average rank.

References

Marden, J. I. (1995). Analyzing and Modeling Rank Data, London: Chapman & Hall. p.41.

See also

Examples

data(city); city

#>       city suburb country
#> 1 2 3  210     22      10
#> 1 3 2   23      4       1
#> 2 1 3  111     45      14
#> 2 3 1    8      4       0
#> 3 1 2  204    299     125
#> 3 2 1   81    126     152

(X <- permutations(3))

#>      [,1] [,2] [,3]
#> [1,]    1    2    3
#> [2,]    1    3    2
#> [3,]    2    1    3
#> [4,]    2    3    1
#> [5,]    3    1    2
#> [6,]    3    2    1

# the average rank can be computed without this function
normalize <- function(x) x / sum(x)
factorial(3) * apply(t(X) %*% city, 2, normalize)

#>          city suburb  country
#> [1,] 2.081633  2.798 2.880795
#> [2,] 1.554160  1.328 1.543046
#> [3,] 2.364207  1.874 1.576159

# the dataset city is really like three datasets; they can be pooled back
# into one via:
rowSums(city)

#> 1 2 3 1 3 2 2 1 3 2 3 1 3 1 2 3 2 1 
#>   242    28   170    12   628   359 

factorial(3) * apply(t(X) %*% rowSums(city), 2, normalize)

#>          [,1]
#> [1,] 2.498263
#> [2,] 1.473245
#> [3,] 2.028492

# the means matrix is used to summarize the data to the means subspace
# which is the subspace of m! spanned by the columns of permutations(m)
# note that when we project onto that subspace, the projection has the
# same average rank vector :
Smaker(3) %*% city # the projections, table 2.8

#>           city     suburb   country
#> [1,] 136.16666   6.333328 -15.33334
#> [2,]  50.16665 -39.166678 -17.00000
#> [3,] 192.16665 128.833323  52.00000
#> [4,]  20.16666  37.833328  48.66666
#> [5,] 162.16666 205.833327 117.66666
#> [6,]  76.16665 160.333323 116.00000

factorial(3) * apply(t(X) %*% Smaker(3) %*% city, 2, normalize)

#>          city suburb  country
#> [1,] 2.081633  2.798 2.880795
#> [2,] 1.554160  1.328 1.543046
#> [3,] 2.364207  1.874 1.576159

# the residuals can be computed by projecting onto the orthogonal complement
(diag(6) - Smaker(3)) %*% city # residuals

#>            city    suburb    country
#> [1,]  73.833341  15.66667  25.333339
#> [2,] -27.166653  43.16668  18.000004
#> [3,] -81.166653 -83.83332 -37.999996
#> [4,] -12.166659 -33.83333 -48.666661
#> [5,]  41.833341  93.16667   7.333339
#> [6,]   4.833347 -34.33332  36.000004

apply(t(X) %*% city, 2, function(x) x / sum(x) * factorial(3)) # average ranks by group

#>          city suburb  country
#> [1,] 2.081633  2.798 2.880795
#> [2,] 1.554160  1.328 1.543046
#> [3,] 2.364207  1.874 1.576159

apply(t(X) %*% rowSums(city), 2, function(x) x / sum(x) * factorial(3)) # average ranks pooled

#>          [,1]
#> [1,] 2.498263
#> [2,] 1.473245
#> [3,] 2.028492