Create the sufficient statistics calculating matrix for approval data

Create the sufficient statistics calculating matrix for approval data

Tmaker(m, k, d)

Arguments

m
the number of objects

k
the number of objects selected

d
the order-effect for the desired matrix (0 to k)

Value

...

See also

Examples

Tmaker(4, 2, 0) # m

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    1    1    1    1

Tmaker(4, 2, 1) # generates how many of each

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    1    0    0    0
#> [2,]    1    0    0    1    1    0
#> [3,]    0    1    0    1    0    1
#> [4,]    0    0    1    0    1    1

Tmaker(4, 2, 2) # gives data (order = subsets(1:4, 2))

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    0    0    0    0    0
#> [2,]    0    1    0    0    0    0
#> [3,]    0    0    1    0    0    0
#> [4,]    0    0    0    1    0    0
#> [5,]    0    0    0    0    1    0
#> [6,]    0    0    0    0    0    1

Tmaker(5, 2, 0)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    1    1    1    1    1    1    1    1    1     1

Tmaker(5, 2, 1)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    1    1    1    1    0    0    0    0    0     0
#> [2,]    1    0    0    0    1    1    1    0    0     0
#> [3,]    0    1    0    0    1    0    0    1    1     0
#> [4,]    0    0    1    0    0    1    0    1    0     1
#> [5,]    0    0    0    1    0    0    1    0    1     1

Tmaker(5, 2, 2)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1    0    0    0    0    0    0    0    0     0
#>  [2,]    0    1    0    0    0    0    0    0    0     0
#>  [3,]    0    0    1    0    0    0    0    0    0     0
#>  [4,]    0    0    0    1    0    0    0    0    0     0
#>  [5,]    0    0    0    0    1    0    0    0    0     0
#>  [6,]    0    0    0    0    0    1    0    0    0     0
#>  [7,]    0    0    0    0    0    0    1    0    0     0
#>  [8,]    0    0    0    0    0    0    0    1    0     0
#>  [9,]    0    0    0    0    0    0    0    0    1     0
#> [10,]    0    0    0    0    0    0    0    0    0     1

Tmaker(4, 3, 0) #

#>      [,1] [,2] [,3] [,4]
#> [1,]    1    1    1    1

Tmaker(4, 3, 1) # subsets(1:4, 3), 1 is in 1, 2, and 3

#>      [,1] [,2] [,3] [,4]
#> [1,]    1    1    1    0
#> [2,]    1    1    0    1
#> [3,]    1    0    1    1
#> [4,]    0    1    1    1

Tmaker(4, 3, 2) # subsets(1:4, 2)

#>      [,1] [,2] [,3] [,4]
#> [1,]    1    1    0    0
#> [2,]    1    0    1    0
#> [3,]    0    1    1    0
#> [4,]    1    0    0    1
#> [5,]    0    1    0    1
#> [6,]    0    0    1    1

Tmaker(4, 3, 3)

#>      [,1] [,2] [,3] [,4]
#> [1,]    1    0    0    0
#> [2,]    0    1    0    0
#> [3,]    0    0    1    0
#> [4,]    0    0    0    1

data(cookie); cookie

#> $freq
#>     Do-Si-Dos, Lemmon Chalet, Samoas  Do-Si-Dos, Lemmon Chalet, Tagalongs 
#>                                    0                                    1 
#> Do-Si-Dos, Lemmon Chalet, Thin Mints   Do-Si-Dos, Lemmon Chalet, Trefoils 
#>                                    1                                    0 
#>         Do-Si-Dos, Samoas, Tagalongs        Do-Si-Dos, Samoas, Thin Mints 
#>                                    6                                   12 
#>          Do-Si-Dos, Samoas, Trefoils     Do-Si-Dos, Tagalongs, Thin Mints 
#>                                    1                                    7 
#>       Do-Si-Dos, Tagalongs, Trefoils      Do-Si-Dos, Thin Mints, Trefoils 
#>                                    1                                    2 
#>     Lemmon Chalet, Samoas, Tagalongs    Lemmon Chalet, Samoas, Thin Mints 
#>                                    2                                   15 
#>      Lemmon Chalet, Samoas, Trefoils Lemmon Chalet, Tagalongs, Thin Mints 
#>                                    0                                    3 
#>   Lemmon Chalet, Tagalongs, Trefoils  Lemmon Chalet, Thin Mints, Trefoils 
#>                                    2                                    8 
#>        Samoas, Tagalongs, Thin Mints          Samoas, Tagalongs, Trefoils 
#>                                   43                                    6 
#>         Samoas, Thin Mints, Trefoils      Tagalongs, Thin Mints, Trefoils 
#>                                   12                                   10 
#> 
#> $raw
#>        [,1]            [,2]            [,3]           
#>   [1,] "Thin Mints"    "Samoas"        "Tagalongs"    
#>   [2,] "Samoas"        "Lemmon Chalet" "Thin Mints"   
#>   [3,] "Samoas"        "Do-Si-Dos"     "Tagalongs"    
#>   [4,] "Thin Mints"    "Tagalongs"     "Do-Si-Dos"    
#>   [5,] "Thin Mints"    "Samoas"        "Tagalongs"    
#>   [6,] "Tagalongs"     "Trefoils"      "Thin Mints"   
#>   [7,] "Do-Si-Dos"     "Thin Mints"    "Samoas"       
#>   [8,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>   [9,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [10,] "Do-Si-Dos"     "Thin Mints"    "Tagalongs"    
#>  [11,] "Tagalongs"     "Thin Mints"    "Samoas"       
#>  [12,] "Tagalongs"     "Thin Mints"    "Samoas"       
#>  [13,] "Do-Si-Dos"     "Samoas"        "Tagalongs"    
#>  [14,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [15,] "Thin Mints"    "Do-Si-Dos"     "Lemmon Chalet"
#>  [16,] "Trefoils"      "Do-Si-Dos"     "Tagalongs"    
#>  [17,] "Trefoils"      "Tagalongs"     "Samoas"       
#>  [18,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [19,] "Tagalongs"     "Samoas"        "Thin Mints"   
#>  [20,] "Lemmon Chalet" "Thin Mints"    "Samoas"       
#>  [21,] "Samoas"        "Tagalongs"     "Thin Mints"   
#>  [22,] "Do-Si-Dos"     "Thin Mints"    "Samoas"       
#>  [23,] "Samoas"        "Thin Mints"    "Trefoils"     
#>  [24,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [25,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [26,] "Tagalongs"     "Lemmon Chalet" "Thin Mints"   
#>  [27,] "Tagalongs"     "Samoas"        "Do-Si-Dos"    
#>  [28,] "Lemmon Chalet" "Samoas"        "Thin Mints"   
#>  [29,] "Samoas"        "Trefoils"      "Thin Mints"   
#>  [30,] "Thin Mints"    "Samoas"        "Trefoils"     
#>  [31,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [32,] "Lemmon Chalet" "Samoas"        "Tagalongs"    
#>  [33,] "Tagalongs"     "Trefoils"      "Lemmon Chalet"
#>  [34,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [35,] "Tagalongs"     "Thin Mints"    "Trefoils"     
#>  [36,] "Samoas"        "Thin Mints"    "Trefoils"     
#>  [37,] "Samoas"        "Trefoils"      "Thin Mints"   
#>  [38,] "Do-Si-Dos"     "Thin Mints"    "Tagalongs"    
#>  [39,] "Lemmon Chalet" "Samoas"        "Thin Mints"   
#>  [40,] "Tagalongs"     "Thin Mints"    "Samoas"       
#>  [41,] "Samoas"        "Do-Si-Dos"     "Thin Mints"   
#>  [42,] "Lemmon Chalet" "Samoas"        "Thin Mints"   
#>  [43,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [44,] "Tagalongs"     "Thin Mints"    "Samoas"       
#>  [45,] "Lemmon Chalet" "Samoas"        "Thin Mints"   
#>  [46,] "Tagalongs"     "Samoas"        "Thin Mints"   
#>  [47,] "Samoas"        "Tagalongs"     "Thin Mints"   
#>  [48,] "Tagalongs"     "Thin Mints"    "Samoas"       
#>  [49,] "Tagalongs"     "Trefoils"      "Thin Mints"   
#>  [50,] "Thin Mints"    "Samoas"        "Do-Si-Dos"    
#>  [51,] "Thin Mints"    "Tagalongs"     "Trefoils"     
#>  [52,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [53,] "Thin Mints"    "Samoas"        "Tagalongs"    
#>  [54,] "Samoas"        "Do-Si-Dos"     "Thin Mints"   
#>  [55,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [56,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [57,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [58,] "Tagalongs"     "Thin Mints"    "Lemmon Chalet"
#>  [59,] "Lemmon Chalet" "Samoas"        "Thin Mints"   
#>  [60,] "Thin Mints"    "Samoas"        "Tagalongs"    
#>  [61,] "Trefoils"      "Thin Mints"    "Samoas"       
#>  [62,] "Samoas"        "Tagalongs"     "Trefoils"     
#>  [63,] "Samoas"        "Trefoils"      "Tagalongs"    
#>  [64,] "Trefoils"      "Do-Si-Dos"     "Samoas"       
#>  [65,] "Lemmon Chalet" "Thin Mints"    "Samoas"       
#>  [66,] "Do-Si-Dos"     "Trefoils"      "Thin Mints"   
#>  [67,] "Trefoils"      "Thin Mints"    "Tagalongs"    
#>  [68,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [69,] "Lemmon Chalet" "Thin Mints"    "Trefoils"     
#>  [70,] "Samoas"        "Thin Mints"    "Trefoils"     
#>  [71,] "Thin Mints"    "Tagalongs"     "Trefoils"     
#>  [72,] "Thin Mints"    "Tagalongs"     "Trefoils"     
#>  [73,] "Trefoils"      "Tagalongs"     "Thin Mints"   
#>  [74,] "Samoas"        "Tagalongs"     "Thin Mints"   
#>  [75,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [76,] "Samoas"        "Lemmon Chalet" "Thin Mints"   
#>  [77,] "Thin Mints"    "Samoas"        "Tagalongs"    
#>  [78,] "Do-Si-Dos"     "Samoas"        "Thin Mints"   
#>  [79,] "Lemmon Chalet" "Tagalongs"     "Thin Mints"   
#>  [80,] "Tagalongs"     "Lemmon Chalet" "Trefoils"     
#>  [81,] "Tagalongs"     "Do-Si-Dos"     "Samoas"       
#>  [82,] "Do-Si-Dos"     "Samoas"        "Thin Mints"   
#>  [83,] "Trefoils"      "Thin Mints"    "Lemmon Chalet"
#>  [84,] "Thin Mints"    "Trefoils"      "Lemmon Chalet"
#>  [85,] "Samoas"        "Tagalongs"     "Do-Si-Dos"    
#>  [86,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [87,] "Trefoils"      "Tagalongs"     "Samoas"       
#>  [88,] "Thin Mints"    "Trefoils"      "Samoas"       
#>  [89,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [90,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [91,] "Tagalongs"     "Thin Mints"    "Trefoils"     
#>  [92,] "Samoas"        "Thin Mints"    "Lemmon Chalet"
#>  [93,] "Thin Mints"    "Tagalongs"     "Samoas"       
#>  [94,] "Lemmon Chalet" "Thin Mints"    "Samoas"       
#>  [95,] "Do-Si-Dos"     "Samoas"        "Tagalongs"    
#>  [96,] "Thin Mints"    "Trefoils"      "Do-Si-Dos"    
#>  [97,] "Samoas"        "Thin Mints"    "Tagalongs"    
#>  [98,] "Do-Si-Dos"     "Tagalongs"     "Lemmon Chalet"
#>  [99,] "Thin Mints"    "Do-Si-Dos"     "Tagalongs"    
#> [100,] "Samoas"        "Thin Mints"    "Do-Si-Dos"    
#> [101,] "Lemmon Chalet" "Thin Mints"    "Trefoils"     
#> [102,] "Thin Mints"    "Do-Si-Dos"     "Tagalongs"    
#> [103,] "Trefoils"      "Lemmon Chalet" "Thin Mints"   
#> [104,] "Thin Mints"    "Samoas"        "Tagalongs"    
#> [105,] "Do-Si-Dos"     "Samoas"        "Thin Mints"   
#> [106,] "Lemmon Chalet" "Thin Mints"    "Trefoils"     
#> [107,] "Lemmon Chalet" "Thin Mints"    "Samoas"       
#> [108,] "Samoas"        "Trefoils"      "Thin Mints"   
#> [109,] "Do-Si-Dos"     "Samoas"        "Thin Mints"   
#> [110,] "Lemmon Chalet" "Trefoils"      "Thin Mints"   
#> [111,] "Thin Mints"    "Lemmon Chalet" "Samoas"       
#> [112,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"   
#> [113,] "Samoas"        "Trefoils"      "Thin Mints"   
#> [114,] "Thin Mints"    "Tagalongs"     "Samoas"       
#> [115,] "Tagalongs"     "Trefoils"      "Thin Mints"   
#> [116,] "Tagalongs"     "Samoas"        "Thin Mints"   
#> [117,] "Samoas"        "Thin Mints"    "Tagalongs"    
#> [118,] "Thin Mints"    "Samoas"        "Trefoils"     
#> [119,] "Tagalongs"     "Lemmon Chalet" "Samoas"       
#> [120,] "Thin Mints"    "Samoas"        "Trefoils"     
#> [121,] "Thin Mints"    "Samoas"        "Tagalongs"    
#> [122,] "Lemmon Chalet" "Thin Mints"    "Samoas"       
#> [123,] "Thin Mints"    "Do-Si-Dos"     "Samoas"       
#> [124,] "Thin Mints"    "Samoas"        "Tagalongs"    
#> [125,] "Thin Mints"    "Tagalongs"     "Samoas"       
#> [126,] "Lemmon Chalet" "Samoas"        "Thin Mints"   
#> [127,] "Trefoils"      "Lemmon Chalet" "Thin Mints"   
#> [128,] "Tagalongs"     "Samoas"        "Trefoils"     
#> [129,] "Samoas"        "Thin Mints"    "Tagalongs"    
#> [130,] "Trefoils"      "Tagalongs"     "Samoas"       
#> [131,] "Samoas"        "Do-Si-Dos"     "Thin Mints"   
#> [132,] "Do-Si-Dos"     "Thin Mints"    "Tagalongs"    
#> 
#> $sorted
#>        [,1]            [,2]            [,3]        
#>   [1,] "Do-Si-Dos"     "Lemmon Chalet" "Tagalongs" 
#>   [2,] "Do-Si-Dos"     "Lemmon Chalet" "Thin Mints"
#>   [3,] "Do-Si-Dos"     "Samoas"        "Tagalongs" 
#>   [4,] "Do-Si-Dos"     "Samoas"        "Tagalongs" 
#>   [5,] "Do-Si-Dos"     "Samoas"        "Tagalongs" 
#>   [6,] "Do-Si-Dos"     "Samoas"        "Tagalongs" 
#>   [7,] "Do-Si-Dos"     "Samoas"        "Tagalongs" 
#>   [8,] "Do-Si-Dos"     "Samoas"        "Tagalongs" 
#>   [9,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [10,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [11,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [12,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [13,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [14,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [15,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [16,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [17,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [18,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [19,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [20,] "Do-Si-Dos"     "Samoas"        "Thin Mints"
#>  [21,] "Do-Si-Dos"     "Samoas"        "Trefoils"  
#>  [22,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [23,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [24,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [25,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [26,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [27,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [28,] "Do-Si-Dos"     "Tagalongs"     "Thin Mints"
#>  [29,] "Do-Si-Dos"     "Tagalongs"     "Trefoils"  
#>  [30,] "Do-Si-Dos"     "Thin Mints"    "Trefoils"  
#>  [31,] "Do-Si-Dos"     "Thin Mints"    "Trefoils"  
#>  [32,] "Lemmon Chalet" "Samoas"        "Tagalongs" 
#>  [33,] "Lemmon Chalet" "Samoas"        "Tagalongs" 
#>  [34,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [35,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [36,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [37,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [38,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [39,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [40,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [41,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [42,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [43,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [44,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [45,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [46,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [47,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [48,] "Lemmon Chalet" "Samoas"        "Thin Mints"
#>  [49,] "Lemmon Chalet" "Tagalongs"     "Thin Mints"
#>  [50,] "Lemmon Chalet" "Tagalongs"     "Thin Mints"
#>  [51,] "Lemmon Chalet" "Tagalongs"     "Thin Mints"
#>  [52,] "Lemmon Chalet" "Tagalongs"     "Trefoils"  
#>  [53,] "Lemmon Chalet" "Tagalongs"     "Trefoils"  
#>  [54,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [55,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [56,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [57,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [58,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [59,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [60,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [61,] "Lemmon Chalet" "Thin Mints"    "Trefoils"  
#>  [62,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [63,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [64,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [65,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [66,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [67,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [68,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [69,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [70,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [71,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [72,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [73,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [74,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [75,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [76,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [77,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [78,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [79,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [80,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [81,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [82,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [83,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [84,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [85,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [86,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [87,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [88,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [89,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [90,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [91,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [92,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [93,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [94,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [95,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [96,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [97,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [98,] "Samoas"        "Tagalongs"     "Thin Mints"
#>  [99,] "Samoas"        "Tagalongs"     "Thin Mints"
#> [100,] "Samoas"        "Tagalongs"     "Thin Mints"
#> [101,] "Samoas"        "Tagalongs"     "Thin Mints"
#> [102,] "Samoas"        "Tagalongs"     "Thin Mints"
#> [103,] "Samoas"        "Tagalongs"     "Thin Mints"
#> [104,] "Samoas"        "Tagalongs"     "Thin Mints"
#> [105,] "Samoas"        "Tagalongs"     "Trefoils"  
#> [106,] "Samoas"        "Tagalongs"     "Trefoils"  
#> [107,] "Samoas"        "Tagalongs"     "Trefoils"  
#> [108,] "Samoas"        "Tagalongs"     "Trefoils"  
#> [109,] "Samoas"        "Tagalongs"     "Trefoils"  
#> [110,] "Samoas"        "Tagalongs"     "Trefoils"  
#> [111,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [112,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [113,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [114,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [115,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [116,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [117,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [118,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [119,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [120,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [121,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [122,] "Samoas"        "Thin Mints"    "Trefoils"  
#> [123,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [124,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [125,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [126,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [127,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [128,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [129,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [130,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [131,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> [132,] "Tagalongs"     "Thin Mints"    "Trefoils"  
#> 
#> $cookies
#> [1] "Do-Si-Dos"     "Lemmon Chalet" "Samoas"        "Tagalongs"    
#> [5] "Thin Mints"    "Trefoils"     
#> 

## voting statistics at different levels
############################################################

# projection onto V0: the number of people in survey
effectsOnV0 <- Tmaker(6, 3, 0) %*% cookie$freq
colnames(effectsOnV0) <- "Total Votes"
effectsOnV0 # = sum(cookie$freq)

#>      Total Votes
#> [1,]         132

# projection onto V1: the number of people voting for each cookie
effectsOnV1 <- Tmaker(6, 3, 1) %*% cookie$freq
row.names(effectsOnV1) <- cookie$cookies
colnames(effectsOnV1) <- "Total Votes"
effectsOnV1

#>               Total Votes
#> Do-Si-Dos              31
#> Lemmon Chalet          32
#> Samoas                 97
#> Tagalongs              81
#> Thin Mints            113
#> Trefoils               42

# projection onto V2: the number of people voting for each cookie-pair
effectsOnV2 <- Tmaker(6, 3, 2) %*% cookie$freq
row.names(effectsOnV2) <- sapply(subsets(cookie$cookies, 2), paste, collapse = ", ")
colnames(effectsOnV2) <- "Total Votes"
effectsOnV2

#>                           Total Votes
#> Do-Si-Dos, Lemmon Chalet            2
#> Do-Si-Dos, Samoas                  19
#> Do-Si-Dos, Tagalongs               15
#> Do-Si-Dos, Thin Mints              22
#> Do-Si-Dos, Trefoils                 4
#> Lemmon Chalet, Samoas              17
#> Lemmon Chalet, Tagalongs            8
#> Lemmon Chalet, Thin Mints          27
#> Lemmon Chalet, Trefoils            10
#> Samoas, Tagalongs                  57
#> Samoas, Thin Mints                 82
#> Samoas, Trefoils                   19
#> Tagalongs, Thin Mints              63
#> Tagalongs, Trefoils                19
#> Thin Mints, Trefoils               32

# projection onto V3: the number of people voting for each cookie-triple
effectsOnV3 <- Tmaker(6, 3, 3) %*% cookie$freq
row.names(effectsOnV3) <- sapply(subsets(cookie$cookies, 3), paste, collapse = ", ")
colnames(effectsOnV3) <- "Total Votes"
effectsOnV3 # = t(t(cookie$freq)) = the (freq) data

#>                                      Total Votes
#> Do-Si-Dos, Lemmon Chalet, Samoas               0
#> Do-Si-Dos, Lemmon Chalet, Tagalongs            1
#> Do-Si-Dos, Lemmon Chalet, Thin Mints           1
#> Do-Si-Dos, Lemmon Chalet, Trefoils             0
#> Do-Si-Dos, Samoas, Tagalongs                   6
#> Do-Si-Dos, Samoas, Thin Mints                 12
#> Do-Si-Dos, Samoas, Trefoils                    1
#> Do-Si-Dos, Tagalongs, Thin Mints               7
#> Do-Si-Dos, Tagalongs, Trefoils                 1
#> Do-Si-Dos, Thin Mints, Trefoils                2
#> Lemmon Chalet, Samoas, Tagalongs               2
#> Lemmon Chalet, Samoas, Thin Mints             15
#> Lemmon Chalet, Samoas, Trefoils                0
#> Lemmon Chalet, Tagalongs, Thin Mints           3
#> Lemmon Chalet, Tagalongs, Trefoils             2
#> Lemmon Chalet, Thin Mints, Trefoils            8
#> Samoas, Tagalongs, Thin Mints                 43
#> Samoas, Tagalongs, Trefoils                    6
#> Samoas, Thin Mints, Trefoils                  12
#> Tagalongs, Thin Mints, Trefoils               10