Count Similarly Margined Contingency Tables

Count the number of contingency tables with the same marginals as a given table.

count_tables(
  table,
  A = hmat(dim(table), as.list(1:length(dim(table)))),
  dir = tempdir(),
  quiet = TRUE,
  cache = TRUE,
  ...
)

Arguments

table
the table of interest

A
the configuration/transpose design matrix

dir
directory to place the files in, without an ending /

quiet
show latte output

cache
use count (default) or fcount

...
arguments to pass to latte_count()

Value

an integer

Details

count_tables uses LattE's count function (via latte's latte_count()) to count the tables. In many cases, the number of such tables is enormous. In these cases, instead of giving back an integer count_tables provides a character string with the integer in it; see examples.

See also

Examples

if (has_latte()) {

data(politics); politics
count_tables(politics)
(A <- hmat(c(2,2), list(1, 2)))
count_tables(politics, A)

data(handy); handy
count_tables(handy)

data(HairEyeColor); HairEyeColor
eyeHairColor <- margin.table(HairEyeColor, 2:1)
count_tables(eyeHairColor)

system.time(count_tables(eyeHairColor)) # it was computed above
system.time(count_tables(eyeHairColor)) # it was computed above
system.time(count_tables(eyeHairColor, cache = FALSE))
system.time(count_tables(eyeHairColor, cache = FALSE))

library(gmp)
as.bigz(count_tables(eyeHairColor))

# notice that even tables with small cells can have huge fibers
data(drugs); drugs
count_tables(drugs)

# 0-1 tables can be very hard and produce very large fibers
# the 4x4 table below has 154 elements in its independence fiber
# the 5x5 has 16830, and the compute times are on the order of
# 1 and 10 seconds, respectively.
# the following block is omited for faster R check time
if (FALSE) {
set.seed(1)
n <- 5
tab <- matrix(sample(0:1, n^2, replace = TRUE), nrow = n)
dimnames(tab) <- list(X = paste0("x", 1:n), Y = paste0("y", 1:n))
tab
count_tables(tab)
}

count_tables(eyeHairColor, quiet = FALSE)

}

#> 
#> Attaching package: ‘gmp’

#> The following objects are masked from ‘package:base’:
#> 
#>     %*%, apply, crossprod, matrix, tcrossprod

#> [-66 286 0 0 17 0 71 127 93 0 0 0 64 0 0 0]
#> [1 -1 0 0 -1 1 0 0 0 0 0 0 0 0 0 0]
#> [-1 0 1 0 1 0 -1 0 0 0 0 0 0 0 0 0]
#> [1 -1 0 0 0 0 0 0 -1 1 0 0 0 0 0 0]
#> [0 0 0 0 1 0 -1 0 -1 0 1 0 0 0 0 0]
#> [0 0 0 0 1 0 0 -1 -1 0 0 1 0 0 0 0]
#> [-1 0 0 1 1 0 0 -1 0 0 0 0 0 0 0 0]
#> [1 -1 0 0 0 0 0 0 0 0 0 0 -1 1 0 0]
#> [0 0 0 0 1 0 -1 0 0 0 0 0 -1 0 1 0]
#> [0 0 0 0 1 0 0 -1 0 0 0 0 -1 0 0 1]
#> Ray Support Algorithm.
#> The codimension is 16
#> 
#>   Left =   7,  Col =  15,  Size =       10  Time:  0.00
#>   Left =   7,  Col =  15,  Size =       12,   Index = 0/4
#>   Left =   7,  Col =  15,  Size =       30,  Time:  0.00                
#> 
#>   Left =   6,  Col =  11,  Size =       30  Time:  0.00
#>   Left =   6,  Col =  11,  Size =       18,   Index = 0/12
#>   Left =   6,  Col =  11,  Size =       51,  Time:  0.00                
#> 
#>   Left =   5,  Col =  14,  Size =       51  Time:  0.00
#>   Left =   5,  Col =  14,  Size =       23,   Index = 0/18
#>   Left =   5,  Col =  14,  Size =       74,  Time:  0.00                
#> 
#>   Left =   4,  Col =  12,  Size =       74  Time:  0.00
#>   Left =   4,  Col =  12,  Size =       31,   Index = 0/28
#>   Left =   4,  Col =  12,  Size =       93,  Time:  0.00                
#> 
#>   Left =   3,  Col =  13,  Size =       93  Time:  0.00
#>   Left =   3,  Col =  13,  Size =       57,   Index = 0/40
#>   Left =   3,  Col =  13,  Size =      196,  Time:  0.00                
#> 
#>   Left =   2,  Col =   7,  Size =      196  Time:  0.00
#>   Left =   2,  Col =   7,  Size =      126,   Index = 0/76
#>   Left =   2,  Col =   7,  Size =      248,  Time:  0.00                
#> 
#>   Left =   1,  Col =  16,  Size =      248  Time:  0.00
#>   Left =   1,  Col =  16,  Size =      248,  Time:  0.00                
#> 1225914276768514

#> This is LattE integrale 1.6
#> Available from http://www.math.ucdavis.edu/~latte/
#> 
#> Invocation: /Applications/latte/bin/count.bin /var/folders/r3/126_d6t55f5d32tplbg5mk1d0c48s9/T//RtmpSvIp4f/2020_03_23_20_05_52_g23Ve69fXR/count_code.latte 
#> Checking whether the input polytope is empty or not...size = 32 x 17
#> Number Type = rational
#> done.
#> Removing redundant inequalities and finding hidden equalities using cddlib...done. 
#> Ax <= b, given as (b|-A):
#> =========================
#> [0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0]
#> [0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0]
#> [0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0]
#> [0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0]
#> [0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0]
#> [0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0]
#> [0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0]
#> [0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0]
#> [0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0]
#> [0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0]
#> [0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0]
#> [0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0]
#> [0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0]
#> [0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0]
#> [0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0]
#> [0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1]
#> 
#> Ax = b, given as (b|-A):
#> ========================
#> [220 -1 -1 -1 -1 0 0 0 0 0 0 0 0 0 0 0 0]
#> [215 0 0 0 0 -1 -1 -1 -1 0 0 0 0 0 0 0 0]
#> [93 0 0 0 0 0 0 0 0 -1 -1 -1 -1 0 0 0 0]
#> [64 0 0 0 0 0 0 0 0 0 0 0 0 -1 -1 -1 -1]
#> [108 -1 0 0 0 -1 0 0 0 -1 0 0 0 -1 0 0 0]
#> [286 0 -1 0 0 0 -1 0 0 0 -1 0 0 0 -1 0 0]
#> [71 0 0 -1 0 0 0 -1 0 0 0 -1 0 0 0 -1 0]
#> 
#> Computing hermitean normal form.
#> sol:
#> [220 215 93 64 -484 -198 -127 0 0 0 0 0 0 0 0 0]
#> Particular solution:
#> Basis:
#> New inequalities:
#> [-66 1 -1 1 0 0 -1 1 0 0 0]
#> [286 -1 0 -1 0 0 0 -1 0 0 0]
#> [0 0 1 0 0 0 0 0 0 0 0]
#> [0 0 0 0 0 0 1 0 0 0 0]
#> [17 -1 1 0 1 1 1 0 1 1 0]
#> [0 1 0 0 0 0 0 0 0 0 0]
#> [71 0 -1 0 -1 0 0 0 -1 0 0]
#> [127 0 0 0 0 -1 -1 0 0 -1 0]
#> [93 0 0 -1 -1 -1 0 0 0 0 0]
#> [0 0 0 1 0 0 0 0 0 0 0]
#> [0 0 0 0 1 0 0 0 0 0 0]
#> [0 0 0 0 0 1 0 0 0 0 0]
#> [64 0 0 0 0 0 0 -1 -1 -1 0]
#> [0 0 0 0 0 0 0 1 0 0 0]
#> [0 0 0 0 0 0 0 0 1 0 0]
#> [0 0 0 0 0 0 0 0 0 1 0]
#> Time for reading and preprocessing: 0.11 sec
#> The polytope has 248 vertices.
#> Time for computing vertices and supporting cones: 0.01 sec
#> Time for dualizing general cones: 0 sec
#> Dualizing all cones...All cones are now dualized.
#> Time for dualizing general cones: 0 sec
#> decomposeCones_Single: Decomposing all cones. (Memory Save on)
#> 248 cones total to be done!decomposeCones_Single: degree = 1
#> Number of cones: 248
#> Triangulating cone... done.
#> 0 vertex cones done. 
#> Triangulating cone... done.
#> 1 vertex cones done. 
#> Triangulating cone... done.
#> 2 vertex cones done. 
#> Triangulating cone... done.
#> 3 vertex cones done. 
#> Triangulating cone... done.
#> 4 vertex cones done. 
#> Triangulating cone... done.
#> 5 vertex cones done. 
#> Triangulating cone... done.
#> 6 vertex cones done. 
#> Triangulating cone... done.
#> 7 vertex cones done. 
#> Triangulating cone... done.
#> 8 vertex cones done. 
#> Triangulating cone... done.
#> 9 vertex cones done. 
#> Triangulating cone... done.
#> 10 vertex cones done. 
#> Triangulating cone... done.
#> 11 vertex cones done. 
#> Triangulating cone... done.
#> 12 vertex cones done. 
#> Triangulating cone... done.
#> 13 vertex cones done. 
#> Triangulating cone... done.
#> 14 vertex cones done. 
#> Triangulating cone... done.
#> 15 vertex cones done. 
#> Triangulating cone... done.
#> 16 vertex cones done. 
#> Triangulating cone... done.
#> 17 vertex cones done. 
#> Triangulating cone... done.
#> 18 vertex cones done. 
#> Triangulating cone... done.
#> 19 vertex cones done. 
#> Triangulating cone... done.
#> 20 vertex cones done. 
#> Triangulating cone... done.
#> 21 vertex cones done. 
#> Triangulating cone... done.
#> 22 vertex cones done. 
#> Triangulating cone... done.
#> 23 vertex cones done. 
#> Triangulating cone... done.
#> 24 vertex cones done. 
#> Triangulating cone... done.
#> 25 vertex cones done. 
#> Triangulating cone... done.
#> 26 vertex cones done. 
#> Triangulating cone... done.
#> 27 vertex cones done. 
#> Triangulating cone... done.
#> 28 vertex cones done. 
#> Triangulating cone... done.
#> 29 vertex cones done. 
#> Triangulating cone... done.
#> 30 vertex cones done. 
#> Triangulating cone... done.
#> 31 vertex cones done. 
#> Triangulating cone... done.
#> 32 vertex cones done. 
#> Triangulating cone... done.
#> 33 vertex cones done. 
#> Triangulating cone... done.
#> 34 vertex cones done. 
#> Triangulating cone... done.
#> 35 vertex cones done. 
#> Triangulating cone... done.
#> 36 vertex cones done. 
#> Triangulating cone... done.
#> 37 vertex cones done. 
#> Triangulating cone... done.
#> 38 vertex cones done. 
#> Triangulating cone... done.
#> 39 vertex cones done. 
#> Triangulating cone... done.
#> 40 vertex cones done. 
#> Triangulating cone... done.
#> 41 vertex cones done. 
#> Triangulating cone... done.
#> 42 vertex cones done. 
#> Triangulating cone... done.
#> 43 vertex cones done. 
#> Triangulating cone... done.
#> 44 vertex cones done. 
#> Triangulating cone... done.
#> 45 vertex cones done. 
#> Triangulating cone... done.
#> 46 vertex cones done. 
#> Triangulating cone... done.
#> 47 vertex cones done. 
#> Triangulating cone... done.
#> 48 vertex cones done. 
#> Triangulating cone... done.
#> 49 vertex cones done. 
#> Triangulating cone... done.
#> 50 vertex cones done. 
#> Triangulating cone... done.
#> 51 vertex cones done. 
#> Triangulating cone... done.
#> 52 vertex cones done. 
#> Triangulating cone... done.
#> 53 vertex cones done. 
#> Triangulating cone... done.
#> 54 vertex cones done. 
#> Triangulating cone... done.
#> 55 vertex cones done. 
#> Triangulating cone... done.
#> 56 vertex cones done. 
#> Triangulating cone... done.
#> 57 vertex cones done. 
#> Triangulating cone... done.
#> 58 vertex cones done. 
#> Triangulating cone... done.
#> 59 vertex cones done. 
#> Triangulating cone... done.
#> 60 vertex cones done. 
#> Triangulating cone... done.
#> 61 vertex cones done. 
#> Triangulating cone... done.
#> 62 vertex cones done. 
#> Triangulating cone... done.
#> 63 vertex cones done. 
#> Triangulating cone... done.
#> 64 vertex cones done. 
#> Triangulating cone... done.
#> 65 vertex cones done. 
#> Triangulating cone... done.
#> 66 vertex cones done. 
#> Triangulating cone... done.
#> 67 vertex cones done. 
#> Triangulating cone... done.
#> 68 vertex cones done. 
#> Triangulating cone... done.
#> 69 vertex cones done. 
#> Triangulating cone... done.
#> 70 vertex cones done. 
#> Triangulating cone... done.
#> 71 vertex cones done. 
#> Triangulating cone... done.
#> 72 vertex cones done. 
#> Triangulating cone... done.
#> 73 vertex cones done. 
#> Triangulating cone... done.
#> 74 vertex cones done. 
#> Triangulating cone... done.
#> 75 vertex cones done. 
#> Triangulating cone... done.
#> 76 vertex cones done. 
#> Triangulating cone... done.
#> 77 vertex cones done. 
#> Triangulating cone... done.
#> 78 vertex cones done. 
#> Triangulating cone... done.
#> 79 vertex cones done. 
#> Triangulating cone... done.
#> 80 vertex cones done. 
#> Triangulating cone... done.
#> 81 vertex cones done. 
#> Triangulating cone... done.
#> 82 vertex cones done. 
#> Triangulating cone... done.
#> 83 vertex cones done. 
#> Triangulating cone... done.
#> 84 vertex cones done. 
#> Triangulating cone... done.
#> 85 vertex cones done. 
#> Triangulating cone... done.
#> 86 vertex cones done. 
#> Triangulating cone... done.
#> 87 vertex cones done. 
#> Triangulating cone... done.
#> 88 vertex cones done. 
#> Triangulating cone... done.
#> 89 vertex cones done. 
#> Triangulating cone... done.
#> 90 vertex cones done. 
#> Triangulating cone... done.
#> 91 vertex cones done. 
#> Triangulating cone... done.
#> 92 vertex cones done. 
#> Triangulating cone... done.
#> 93 vertex cones done. 
#> Triangulating cone... done.
#> 94 vertex cones done. 
#> Triangulating cone... done.
#> 95 vertex cones done. 
#> Triangulating cone... done.
#> 96 vertex cones done. 
#> Triangulating cone... done.
#> 97 vertex cones done. 
#> Triangulating cone... done.
#> 98 vertex cones done. 
#> Triangulating cone... done.
#> 99 vertex cones done. 
#> Triangulating cone... done.
#> 100 vertex cones done. 
#> Triangulating cone... done.
#> 101 vertex cones done. 
#> Triangulating cone... done.
#> 102 vertex cones done. 
#> Triangulating cone... done.
#> 103 vertex cones done. 
#> Triangulating cone... done.
#> 104 vertex cones done. 
#> Triangulating cone... done.
#> 105 vertex cones done. 
#> Triangulating cone... done.
#> 106 vertex cones done. 
#> Triangulating cone... done.
#> 107 vertex cones done. 
#> Triangulating cone... done.
#> 108 vertex cones done. 
#> Triangulating cone... done.
#> 109 vertex cones done. 
#> Triangulating cone... done.
#> 110 vertex cones done. 
#> Triangulating cone... done.
#> 111 vertex cones done. 
#> Triangulating cone... done.
#> 112 vertex cones done. 
#> Triangulating cone... done.
#> 113 vertex cones done. 
#> Triangulating cone... done.
#> 114 vertex cones done. 
#> Triangulating cone... done.
#> 115 vertex cones done. 
#> Triangulating cone... done.
#> 116 vertex cones done. 
#> Triangulating cone... done.
#> 117 vertex cones done. 
#> Triangulating cone... done.
#> 118 vertex cones done. 
#> Triangulating cone... done.
#> 119 vertex cones done. 
#> Triangulating cone... done.
#> 120 vertex cones done. 
#> Triangulating cone... done.
#> 121 vertex cones done. 
#> Triangulating cone... done.
#> 122 vertex cones done. 
#> Triangulating cone... done.
#> 123 vertex cones done. 
#> Triangulating cone... done.
#> 124 vertex cones done. 
#> Triangulating cone... done.
#> 125 vertex cones done. 
#> Triangulating cone... done.
#> 126 vertex cones done. 
#> Triangulating cone... done.
#> 127 vertex cones done. 
#> Triangulating cone... done.
#> 128 vertex cones done. 
#> Triangulating cone... done.
#> 129 vertex cones done. 
#> Triangulating cone... done.
#> 130 vertex cones done. 
#> Triangulating cone... done.
#> 131 vertex cones done. 
#> Triangulating cone... done.
#> 132 vertex cones done. 
#> Triangulating cone... done.
#> 133 vertex cones done. 
#> Triangulating cone... done.
#> 134 vertex cones done. 
#> Triangulating cone... done.
#> 135 vertex cones done. 
#> Triangulating cone... done.
#> 136 vertex cones done. 
#> Triangulating cone... done.
#> 137 vertex cones done. 
#> Triangulating cone... done.
#> 138 vertex cones done. 
#> Triangulating cone... done.
#> 139 vertex cones done. 
#> Triangulating cone... done.
#> 140 vertex cones done. 
#> Triangulating cone... done.
#> 141 vertex cones done. 
#> Triangulating cone... done.
#> 142 vertex cones done. 
#> Triangulating cone... done.
#> 143 vertex cones done. 
#> Triangulating cone... done.
#> 144 vertex cones done. 
#> Triangulating cone... done.
#> 145 vertex cones done. 
#> Triangulating cone... done.
#> 146 vertex cones done. 
#> Triangulating cone... done.
#> 147 vertex cones done. 
#> Triangulating cone... done.
#> 148 vertex cones done. 
#> Triangulating cone... done.
#> 149 vertex cones done. 
#> Triangulating cone... done.
#> 150 vertex cones done. 
#> Triangulating cone... done.
#> 151 vertex cones done. 
#> Triangulating cone... done.
#> 152 vertex cones done. 
#> Triangulating cone... done.
#> 153 vertex cones done. 
#> Triangulating cone... done.
#> 154 vertex cones done. 
#> Triangulating cone... done.
#> 155 vertex cones done. 
#> Triangulating cone... done.
#> 156 vertex cones done. 
#> Triangulating cone... done.
#> 157 vertex cones done. 
#> Triangulating cone... done.
#> 158 vertex cones done. 
#> Triangulating cone... done.
#> 159 vertex cones done. 
#> Triangulating cone... done.
#> 160 vertex cones done. 
#> Triangulating cone... done.
#> 161 vertex cones done. 
#> Triangulating cone... done.
#> 162 vertex cones done. 
#> Triangulating cone... done.
#> 163 vertex cones done. 
#> Triangulating cone... done.
#> 164 vertex cones done. 
#> Triangulating cone... done.
#> 165 vertex cones done. 
#> Triangulating cone... done.
#> 166 vertex cones done. 
#> Triangulating cone... done.
#> 167 vertex cones done. 
#> Triangulating cone... done.
#> 168 vertex cones done. 
#> Triangulating cone... done.
#> 169 vertex cones done. 
#> Triangulating cone... done.
#> 170 vertex cones done. 
#> Triangulating cone... done.
#> 171 vertex cones done. 
#> Triangulating cone... done.
#> 172 vertex cones done. 
#> Triangulating cone... done.
#> 173 vertex cones done. 
#> Triangulating cone... done.
#> 174 vertex cones done. 
#> Triangulating cone... done.
#> 175 vertex cones done. 
#> Triangulating cone... done.
#> 176 vertex cones done. 
#> Triangulating cone... done.
#> 177 vertex cones done. 
#> Triangulating cone... done.
#> 178 vertex cones done. 
#> Triangulating cone... done.
#> 179 vertex cones done. 
#> Triangulating cone... done.
#> 180 vertex cones done. 
#> Triangulating cone... done.
#> 181 vertex cones done. 
#> Triangulating cone... done.
#> 182 vertex cones done. 
#> Triangulating cone... done.
#> 183 vertex cones done. 
#> Triangulating cone... done.
#> 184 vertex cones done. 
#> Triangulating cone... done.
#> 185 vertex cones done. 
#> Triangulating cone... done.
#> 186 vertex cones done. 
#> Triangulating cone... done.
#> 187 vertex cones done. 
#> Triangulating cone... done.
#> 188 vertex cones done. 
#> Triangulating cone... done.
#> 189 vertex cones done. 
#> Triangulating cone... done.
#> 190 vertex cones done. 
#> Triangulating cone... done.
#> 191 vertex cones done. 
#> Triangulating cone... done.
#> 192 vertex cones done. 
#> Triangulating cone... done.
#> 193 vertex cones done. 
#> Triangulating cone... done.
#> 194 vertex cones done. 
#> Triangulating cone... done.
#> 195 vertex cones done. 
#> Triangulating cone... done.
#> 196 vertex cones done. 
#> Triangulating cone... done.
#> 197 vertex cones done. 
#> Triangulating cone... done.
#> 198 vertex cones done. 
#> Triangulating cone... done.
#> 199 vertex cones done. 
#> Triangulating cone... done.
#> 200 vertex cones done. 
#> Triangulating cone... done.
#> 201 vertex cones done. 
#> Triangulating cone... done.
#> 202 vertex cones done. 
#> Triangulating cone... done.
#> 203 vertex cones done. 
#> Triangulating cone... done.
#> 204 vertex cones done. 
#> Triangulating cone... done.
#> 205 vertex cones done. 
#> Triangulating cone... done.
#> 206 vertex cones done. 
#> Triangulating cone... done.
#> 207 vertex cones done. 
#> Triangulating cone... done.
#> 208 vertex cones done. 
#> Triangulating cone... done.
#> 209 vertex cones done. 
#> Triangulating cone... done.
#> 210 vertex cones done. 
#> Triangulating cone... done.
#> 211 vertex cones done. 
#> Triangulating cone... done.
#> 212 vertex cones done. 
#> Triangulating cone... done.
#> 213 vertex cones done. 
#> Triangulating cone... done.
#> 214 vertex cones done. 
#> Triangulating cone... done.
#> 215 vertex cones done. 
#> Triangulating cone... done.
#> 216 vertex cones done. 
#> Triangulating cone... done.
#> 217 vertex cones done. 
#> Triangulating cone... done.
#> 218 vertex cones done. 
#> Triangulating cone... done.
#> 219 vertex cones done. 
#> Triangulating cone... done.
#> 220 vertex cones done. 
#> Triangulating cone... done.
#> 221 vertex cones done. 
#> Triangulating cone... done.
#> 222 vertex cones done. 
#> Triangulating cone... done.
#> 223 vertex cones done. 
#> Triangulating cone... done.
#> 224 vertex cones done. 
#> Triangulating cone... done.
#> 225 vertex cones done. 
#> Triangulating cone... done.
#> 226 vertex cones done. 
#> Triangulating cone... done.
#> 227 vertex cones done. 
#> Triangulating cone... done.
#> 228 vertex cones done. 
#> Triangulating cone... done.
#> 229 vertex cones done. 
#> Triangulating cone... done.
#> 230 vertex cones done. 
#> Triangulating cone... done.
#> 231 vertex cones done. 
#> Triangulating cone... done.
#> 232 vertex cones done. 
#> Triangulating cone... done.
#> 233 vertex cones done. 
#> Triangulating cone... done.
#> 234 vertex cones done. 
#> Triangulating cone... done.
#> 235 vertex cones done. 
#> Triangulating cone... done.
#> 236 vertex cones done. 
#> Triangulating cone... done.
#> 237 vertex cones done. 
#> Triangulating cone... done.
#> 238 vertex cones done. 
#> Triangulating cone... done.
#> 239 vertex cones done. 
#> Triangulating cone... done.
#> 240 vertex cones done. 
#> Triangulating cone... done.
#> 241 vertex cones done. 
#> Triangulating cone... done.
#> 242 vertex cones done. 
#> Triangulating cone... done.
#> 243 vertex cones done. 
#> Triangulating cone... done.
#> 244 vertex cones done. 
#> Triangulating cone... done.
#> 245 vertex cones done. 
#> Triangulating cone... done.
#> 246 vertex cones done. 
#> Triangulating cone... done.
#> 247 vertex cones done. 
#> 
#> Total Unimodular Cones: 248
#> Maximum number of simplicial cones in memory at once: 0
#> 
#> ****  The number of lattice points is:   ****
#> Total time: 0.24 sec

#> Time for reading and preprocessing: 0.11 sec
#> Time for computing vertices and supporting cones: 0.01 sec
#> Time for irrationalizing general cones: 0 sec
#> Time for dualizing general cones: 0 sec
#> Time for triangulating cones into simplicial cones: 0 sec
#> Time for Barvinok decomposition and residue calculation: 0.12 sec
#> Total time: 0.24 sec
#> Total number of simplicial cones: 248
#> Total number of unimodular cones: 248
#> Maximum depth of the decomposition tree: 0
#> 

#> [1] "1225914276768514"