Use of Alcohol, Cigarettes, and Marijuana by High School Students

A multi-way contingency table of alcohol, cigarettes, and marijuana use by high school students.

data(drugs)

Format

A 2x2x2 (contingency) table

References

Agresti, A. (2002). Categorical Data Analysis, Basel: John Wiley & Sons, 2ed. (p.322)