Fit a hierarchical log-linear model with algebraic methods

loglinear() fits a hierarchical log-linear model to a dataset (typically a contingency table) and performs an exact conditional test on the fitted model using various distance metrics. The exact test, which is a goodness-of-fit test, is performed via Monte-Carlo sampling from the conditional distribution of the table given the sufficient statistics of the model. In short, inference is drawn by comparing the statistic of the observed table (be it an unnormalized log-likelihood, a Pearson chi-squared value, or another metric) to those of the samples. The proportion of sampled tables with equal to or more extreme values than the observed table is the resulting p-value.

loglinear(
  model,
  data,
  init = tab2vec(data),
  iter = 10000,
  burn = 1000,
  thin = 10,
  engine = c("C++", "R"),
  method = c("ipf", "mcmc"),
  moves,
  ...
)

Arguments

model
hierarchical log-linear model specification

data
data, typically as a table but can be in different formats. see teshape()

init
the initialization of the chain. by default, this is the observed table

iter
number of chain iterations

burn
burn-in

thin
thinning

engine
"C++" or "R"? (C++ is significantly faster)

method
should the expected value (exp) be fit using iterative proportional fitting (via loglin) or the MCMC as the average of the steps?

moves
the markov moves for the mcmc (as columns of a matrix).

...
...

Value

a list containing named elements

steps: an integer matrix whose columns represent individual samples from the mcmc.

moves: the moves used for the proposal distribution in the mcmc, computed with 4ti2 (note that only the positive moves are given).

accept_prob: the empirical transition probability of the moves, including the thinned moves.

param: the fitted parameters of the log linear model.

df: parameters per term in the model

quality: model selection statistics AIC, AICc, and BIC.

residuals: the (unstandardized) pearson residuals (O - E) / sqrt(E)

call: the call.

obs: the contingency table given.

exp: the fit contingency table as an integer array.

A: the sufficient statistics computing matrix (from Tmaker).

p.value: the exact p-values of individual tests, accurate to Monte-Carlo error. these are computed as the proportion of samples with statistics equal to or larger than the oberved statistic.

mid.p.value: the mid p.values, see Agresti pp.20--21.

statistic: the pearson's chi-squared (X2), likelihood ratio (G2), Freeman-Tukey (FT), Cressie-Read (CR), and Neyman modified chi-squared (NM) statistics computed for the table given.

sampsStats: the statistics computed for each mcmc sample.

cells: the number of cells in the table.

method: the method used to estimate the table.

Details

In many ways, loglinear() is like stats::loglin() or MASS::loglm(); however, there are a few key differences.

The first difference is methodological. The tests conducted with loglinear() are exact conditional tests based on the conditional distribution of the data given the sufficient statistics for the model. In other words, they are analogues of Fisher's exact test for generic log-linear models. These tests are made possible by advances in algebraic statistics; see references 1--3 below.

The second difference between loglinear() and loglin() or loglm() is that inference is made through Monte Carlo simulation. In particular, loglinear() leverages Markov moves to sample from the conditional distribution of the data given its sufficent statistics. If the software 4ti2 is installed on your machine, you can use markov() (or let loglinear() use markov()) to generate a Markov basis to use for the Markov moves. This basis is guaranteed to produce a MCMC routine that converges to the conditional distribution of interest. Since loglinear() uses Monte Carlo simulation to conduct inference, and since it uses MCMC to do so, concerns typical to MCMC should be addressed. In particular, issues such as burn in and mixing (autocorrelation of samples) should be addressed. The examples illustrate some of these topics. The result is a p-value that is generated by Monte Carlo simulation. Its standard error is provided (computed as in the standard CLT confidence interval) to give a sense of the Monte Carlo error.

A third way that loglinear() differs from loglin() or loglm() is in generalizing the kinds of tests performed. While those allow for asymptotic unconditional testing using Pearson's X^2 test and the likelihood ratio test, loglinear() gives several test statistics: Pearson's X^2, the likelihood ratio G^2, Freeman-Tukey, Cressie-Read (lambda = 2/3), and Neyman's modified X^2., see the last reference. In other words, to compute the exact p-value, iter = 1e4 samples are sampled from the conditional distribution of the table given the sufficient statistics, and then the proportion of tables that have X^2, G^2, etc. values greater than or equal to that of the observed table is the p value for the (conditional) exact test. A similar, and perhaps preferable approach, simply adds up the probabilities of the tables that have probabilities less than or equal to that of the observed table; this is the first line output in hierarchical and does not use a test statistic.

Some authors (see the third reference) suggest that for discrete problems, a "mid p value" is preferable to the traditional p value, and when presented should be interepreted in the same way. If the p value is defined to be P(samps >= obs), the mid p value is defined to be P(samps > obs) + P(samps == obs)/2. The mid p value is computed for each test.

References

Diaconis, P. and B. Sturmfels (1998). Algebraic Algorithms for Sampling from Conditional Distributions. The Annals of Statistics 26(1), pp.363-397.

Drton, M., B. Sturmfels, and S. Sullivant (2009). Lectures on Algebraic Statistics, Basel: Birkhauser Verlag AG.

Aoki, S., H. Hara, and A. Takemura (2012). Markov Bases in Algebraic Statistics, Springer.

Agresti, A. (2002). Categorical Data Analysis, Basel: John Wiley & Sons, 2ed.

Agresti, A. (1992). A Survey of Exact Inference for Contingency Tables Statistical Science 7(1), pp.131-153.

Read, T. and Cressie, N. (1998). Goodness-of-Fit Statistics for Discrete Multivariate Data, Springer-Verlag.

See also

Examples