Projection onto a variety

A R-based implementation of the gradient descent homotopies

project_onto_variety(
  x0,
  poly,
  dt = 0.05,
  varorder = vars(poly),
  n_correct = 2,
  al = rnorm(length(x0)),
  message = FALSE,
  tol = .Machine$double.eps^(1/4)
)

project_onto_variety_lagrange(
  x0,
  poly,
  method = "newton",
  maxit = 1000,
  tol = .Machine$double.eps^(1/4),
  message = FALSE,
  ...
)

project_onto_variety_gradient_descent(
  x0,
  poly,
  ga = 1e-04,
  max_ga = 0.1,
  method = c("line", "optimal", "fixed"),
  tol = .Machine$double.eps^(1/4),
  maxit = 1000,
  message = FALSE
)

Arguments

x0
Atomic vector, the point to be projected.

poly
An mpoly object, typically created with mp().

dt
The t-mesh size for the homotopy.

varorder
A character vector specifying the variable order to pass to mpoly::as.function.mpoly().

n_correct
The number of Newton correction iterations to use.

al
A numeric vector of length 2; the patch to do projective calculations over.

message
If TRUE, the user is issued messages on the algorithm's progress.

tol
A tolerance; a warning is issued if the magnitude of the residual is larger than tol.

method
Used in project_onto_variety_lagrange() and project_onto_variety_gradient_descent(). In the former, if "newton", a simple R implementation of Newton's method to solve the nonlinear algebraic system generated by the Lagrangian; otherwise, a character string to pass to optim() to minimize the sum of the squares of the Lagrangian. In the latter, the method of selecting the learning rate, "line" (line search) or "fixed".

maxit
Number of maximum iterations in solving Newton's method.

...
Additional arguments to pass to optim() when method is not "newton".

ga
Learning rate for gradient descent.

max_ga
Maximum learning rate for gradient descent line search.

Value

A numeric vector the same length as x0.

References

Griffin, Z. and J. Hauenstein (2015). Real solutions to systems of polynomial equations and parameter continuation. Advances in Geometry 15(2), pp.173--187.

Bates, D., J. Hauenstein, A. Sommese, and C. Wampler (2013). Numerically Solving Polynomial Systems with Bertini. SIAM. pp.34--35.

Examples