Sample from the fiber of a contingency table

Sample from the fiber of a contingency table

rfiber(
  n,
  ...,
  A,
  b,
  x,
  method = c("sis", "walk", "hybrid"),
  dist = c("uniform", "hypergeometric"),
  parallel = FALSE,
  includeRejections = FALSE,
  thin = 1,
  format = c("mat", "vec", "tab"),
  dim
)

Arguments

n
the number of observations

...
...

A
the configuration matrix of the model defining the fiber

b
integer vector; the vector of sufficient statistics

x
integer vector; a vector in the fiber to be used as an initialization point for the "walk" method

method
"sis", "walk", or "hybrid"

dist
"uniform" or "hypergeometric"

parallel
parallelize the sampling?

includeRejections
should rejections be returned? (note: rejection tables aren't completed in the SIS procedure.)

thin
thinning argument (if using walk method)

format
format of the returned moves, "mat", "vec", or "tab"

dim
the dimensions of the table if "tab" is used, see vec2tab()

Value

a named numeric vector

Examples

A <- hmat(c(2,2), 1:2)
b <- rep.int(4, 4)

rfiber(10, A = A, b = b)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    0    4    4    0    0    4    4    1    1     0
#> [2,]    4    0    0    4    4    0    0    3    3     4
#> [3,]    4    0    0    4    4    0    0    3    3     4
#> [4,]    0    4    4    0    0    4    4    1    1     0

rfiber(10, A = A, b = b, format = "vec")

#> [[1]]
#> [1] 3 1 1 3
#> 
#> [[2]]
#> [1] 0 4 4 0
#> 
#> [[3]]
#> [1] 3 1 1 3
#> 
#> [[4]]
#> [1] 2 2 2 2
#> 
#> [[5]]
#> [1] 1 3 3 1
#> 
#> [[6]]
#> [1] 1 3 3 1
#> 
#> [[7]]
#> [1] 3 1 1 3
#> 
#> [[8]]
#> [1] 3 1 1 3
#> 
#> [[9]]
#> [1] 3 1 1 3
#> 
#> [[10]]
#> [1] 1 3 3 1
#> 

rfiber(10, A = A, b = b, format = "tab", dim = c(2, 2))

#> [[1]]
#>      [,1] [,2]
#> [1,]    3    1
#> [2,]    1    3
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    0    4
#> [2,]    4    0
#> 
#> [[3]]
#>      [,1] [,2]
#> [1,]    0    4
#> [2,]    4    0
#> 
#> [[4]]
#>      [,1] [,2]
#> [1,]    3    1
#> [2,]    1    3
#> 
#> [[5]]
#>      [,1] [,2]
#> [1,]    0    4
#> [2,]    4    0
#> 
#> [[6]]
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    3    1
#> 
#> [[7]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    2
#> 
#> [[8]]
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    3    1
#> 
#> [[9]]
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    3    1
#> 
#> [[10]]
#>      [,1] [,2]
#> [1,]    4    0
#> [2,]    0    4
#> 

set.seed(1)
(tab <- rfiber(1, A = A, b = b, format = "tab", dim = c(2, 2))[[1]])

#>      [,1] [,2]
#> [1,]    0    4
#> [2,]    4    0

x <- tab2vec(tab)

if (FALSE)  # requires 4ti2

rfiber(100, A = A, x = x, method = "walk")

library(microbenchmark)
microbenchmark(
  rfiber(100, A = A, b = b),
  rfiber(100, A = A, x = x, method = "walk")
)

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Warning: WARNING: Overflow detection is not available on this architecture/compiler.
#> WARNING: To guarantee correct results, run 4ti2 with arbitrary precision
#> WARNING: by using the option `-parb'

#> Unit: milliseconds
#>                                        expr      min        lq     mean
#>                   rfiber(100, A = A, b = b) 207.4578 215.27040 238.1602
#>  rfiber(100, A = A, x = x, method = "walk")  35.9164  39.30078  43.1260
#>     median        uq       max neval cld
#>  226.68974 246.48116 373.66510   100   b
#>   41.71804  45.45412  64.59009   100  a 

# the distribution of the samples is seemingly near-uniform
tabs <- rfiber(1e4, A, b, format = "vec", parallel = TRUE)

#> Error in is.wholenumber(A): argument "A" is missing, with no default

table(sapply(tabs, paste, collapse = " "))

#> Error in lapply(X = X, FUN = FUN, ...): object 'tabs' not found

tabs <- rfiber(1000, A = A, b = b, format = "vec")
unique(tabs)

#> [[1]]
#> [1] 0 4 4 0
#> 
#> [[2]]
#> [1] 2 2 2 2
#> 
#> [[3]]
#> [1] 4 0 0 4
#> 
#> [[4]]
#> [1] 1 3 3 1
#> 
#> [[5]]
#> [1] 3 1 1 3
#> 

lapply(unique(tabs), vec2tab, dim = c(2, 2))

#> [[1]]
#>      [,1] [,2]
#> [1,]    0    4
#> [2,]    4    0
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]    2    2
#> [2,]    2    2
#> 
#> [[3]]
#>      [,1] [,2]
#> [1,]    4    0
#> [2,]    0    4
#> 
#> [[4]]
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    3    1
#> 
#> [[5]]
#>      [,1] [,2]
#> [1,]    3    1
#> [2,]    1    3
#> 

table(sapply(tabs, paste, collapse = " " ))

#> 
#> 0 4 4 0 1 3 3 1 2 2 2 2 3 1 1 3 4 0 0 4 
#>     202     186     201     212     199 

data(politics); politics

#>            Party
#> Personality Democrat Republican
#>   Introvert        3          7
#>   Extrovert        6          4

tab2vec(politics)

#> 1,1 1,2 2,1 2,2 
#>   3   7   6   4 

b <- A %*% tab2vec(politics)
tabs <- rfiber(1000, A = A, b = b)
unique(t(tabs))

#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    9    8    2
#>  [2,]    3    7    6    4
#>  [3,]    4    6    5    5
#>  [4,]    9    1    0   10
#>  [5,]    0   10    9    1
#>  [6,]    6    4    3    7
#>  [7,]    2    8    7    3
#>  [8,]    7    3    2    8
#>  [9,]    5    5    4    6
#> [10,]    8    2    1    9

table(apply(tabs, 2, paste, collapse = " " )) # roughly uniform

#> 
#> 0 10 9 1  1 9 8 2  2 8 7 3  3 7 6 4  4 6 5 5  5 5 4 6  6 4 3 7  7 3 2 8 
#>       88       97      105       99      106      103      112      106 
#>  8 2 1 9 9 1 0 10 
#>       92       92 

# poisson regression example
J <- 5
A <- rbind(1L, 1:J)
b <- c(5, 15)

system.time(rfiber(1e4, A = A, b = b))

#>    user  system elapsed 
#>  90.613  35.112 128.316 

system.time(rfiber(1e4, A = A, b = b, parallel = TRUE))

#>    user  system elapsed 
#> 109.616  46.618  28.163