The Variety Normal Distribution

Un-normalized density function and random generation for the variety normal distribution with mean equal to poly and "standard deviation" equal to sd. Please see details for caveats.

rvnorm(
  n,
  poly,
  sd,
  output = "simple",
  chains = 4L,
  warmup = floor(n/2),
  keep_warmup = FALSE,
  thin = 1L,
  inject_direct = FALSE,
  verbose = FALSE,
  cores = getOption("mc.cores", 1L),
  normalized = TRUE,
  w,
  vars,
  numerator,
  denominator,
  refresh,
  code_only = FALSE,
  ...
)

Arguments

n
The number of draws desired from each chain after warmup.

poly
An mpoly object.

sd
The "standard deviation" component of the normal kernel.

output
"simple", "more", "stanfit".

chains
The number of chains to run for the random number generation, see stan().

warmup
Number of warmup iterations in stan().

keep_warmup
If TRUE, the MCMC warmup steps are included in the output.

thin
stan() thin parameter.

inject_direct
Directly specify printed polynomial to string inject into the stan code. Requires you specify vars, numerator, and denominator.

verbose
TRUE or FALSE; determines level of messaging.

cores
The number of CPU cores to distribute the chains across, see stan().

normalized
If TRUE, the polynomial is gradient-normalized. This is highly recommended.

w
A named list of box constraints for vectors to be passed to Stan, see examples. A If a single number, a box window (-w,w) is applied to all variables.

vars
A character vector of the indeterminates in the distribution.

numerator, denominator
A character(1) containing the printed numerator of the variety normal distribution.

refresh
The refresh argument of stan(), which governs how much information is provided to the user while sampling.

code_only
If TRUE, will only formulate and return Stan code.

...
Additional parameters to pass to stan().

Value

Either (1) matrix whose rows are the individual draws from the distribution, (2) a tbl_df object with the draws along with additional information, or (3) an object of class stanfit.

Details

If the variety you are interested in is connected, this strategy should work well out of the box. If it isn't, you'll likely need to rely on running multiple chains, and it is very likely, if not probable, that the sampling will be biased to one or more of those components and down-sample others. Question: what is the relative likelihood of each component, or an equal unit of length, on different components? How does this generalize to more varieties of varying dimensions?

Examples